Update java test

+++ modules/java_api/src/test/java/org/intel/openvino/CompiledModelTests.java
@@ -34,7 +34,7 @@ public class CompiledModelTests extends OVTest {
         List<Output> outputs = model.outputs();
 
         assertEquals("fc_out", outputs.get(0).get_any_name());
-        assertEquals(ElementType.f32, outputs.get(0).get_element_type());
+        assertEquals(ElementType.f16, outputs.get(0).get_element_type());
 
         int[] shape = new int[] {1, 10};
         assertArrayEquals("Shape", shape, outputs.get(0).get_shape());
